requirements.txt
albumentations==1.1.0
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cabocha-python==0.69.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
fastprogress==1.0.0
filelock==3.3.2
Flask==2.0.2
fugashi==1.1.1
huggingface-hub==0.1.0
idna==3.3
imageio==2.10.4
importlib-resources==5.4.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
japanize-matplotlib==1.1.3
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.2.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mecab-python3==1.0.4
mistune==0.8.4
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.3
opencv-python==4.5.4.58
opencv-python-headless==4.5.4.58
packaging==21.2
pandas==1.3.4
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pluggy==1.0.0
progress-bar==8
prometheus-client==0.12.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
py==1.10.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
qudida==0.0.4
regex==2021.11.2
requests==2.26.0
sacremoses==0.0.46
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tokenizers==0.10.3
toml==0.10.2
torch==1.10.0
torchinfo==1.5.3
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
zipp==3.6.0
